Updating GravityForms core

+++ docroot/content/plugins/gravityforms/includes/api.php
@@ -42,8 +42,12 @@ class GFAPI {
 			return false;
 		}
 
+		$form_info = GFFormsModel::get_form( $form_id, true );
+		if ( ! $form_info ) {
+			return false;
+		}
+
 		// Loading form columns into meta.
-		$form_info            = GFFormsModel::get_form( $form_id, true );
 		$form['is_active']    = $form_info->is_active;
 		$form['date_created'] = $form_info->date_created;
 		$form['is_trash']     = $form_info->is_trash;
@@ -327,7 +331,7 @@ class GFAPI {
 			$value = sprintf( "'%s'", $value );
 		}
 		$in_str_arr = array_fill( 0, count( $form_ids ), '%d' );
-		$in_str     = join( $in_str_arr, ',' );
+		$in_str     = join( ',', $in_str_arr );
 		$result     = $wpdb->query(
 			$wpdb->prepare(
 				"
@@ -451,20 +455,10 @@ class GFAPI {
 		// Add default confirmation if form has no confirmations.
 		if ( ! isset( $form_meta['confirmations'] ) || empty( $form_meta['confirmations'] ) ) {
 
-			// Generate confirmation ID.
-			$confirmation_id = uniqid();
+			$confirmation = GFFormsModel::get_default_confirmation();
 
 			// Add default confirmation to form.
-			$form_meta['confirmations'][ $confirmation_id ] = array(
-				'id'          => $confirmation_id,
-				'name'        => __( 'Default Confirmation', 'gravityforms' ),
-				'isDefault'   => true,
-				'type'        => 'message',
-				'message'     => __( 'Thanks for contacting us! We will get in touch with you shortly.', 'gravityforms' ),
-				'url'         => '',
-				'pageId'      => '',
-				'queryString' => '',
-			);
+			$form_meta['confirmations'] = array( $confirmation['id'] => $confirmation );
 
 		}
 
@@ -885,10 +879,8 @@ class GFAPI {
 		}
 		$source_url = $entry['source_url'];
 
-		if ( ! isset( $entry['user_agent'] ) ) {
-			$entry['user_agent'] = 'API';
-		}
-		$user_agent = $entry['user_agent'];
+		$entry['user_agent'] = isset( $entry['user_agent'] ) ? sanitize_text_field( $entry['user_agent'] ) : 'API';
+		$user_agent          = $entry['user_agent'];
 
 		if ( empty( $entry['currency'] ) ) {
 			$entry['currency'] = GFCommon::get_currency();
@@ -996,8 +988,7 @@ class GFAPI {
 
 		foreach ( $form['fields'] as $field ) {
 			/* @var GF_Field $field */
-			$type = GFFormsModel::get_input_type( $field );
-			if ( in_array( $type, array( 'html', 'page', 'section' ) ) ) {
+			if ( $field->displayOnly ) {
 				continue;
 			}
 
@@ -1218,7 +1209,7 @@ class GFAPI {
 		$request_ip     = rgars( $form, 'personalData/preventIP' ) ? '' : GFFormsModel::get_ip();
 		$ip             = isset( $entry['ip'] ) ? $entry['ip'] : $request_ip;
 		$source_url     = isset( $entry['source_url'] ) ? $entry['source_url'] : esc_url_raw( GFFormsModel::get_current_page_url() );
-		$user_agent     = isset( $entry['user_agent'] ) ? $entry['user_agent'] : 'API';
+		$user_agent     = isset( $entry['user_agent'] ) ? sanitize_text_field( $entry['user_agent'] ) : 'API';
 		$currency       = isset( $entry['currency'] ) ? $entry['currency'] : GFCommon::get_currency();
 		$payment_status = isset( $entry['payment_status'] ) ? sprintf( "'%s'", esc_sql( $entry['payment_status'] ) ) : 'NULL';
 		$payment_date   = strtotime( rgar( $entry, 'payment_date' ) ) ? sprintf( "'%s'", gmdate( 'Y-m-d H:i:s', strtotime( "{$entry['payment_date']}" ) ) ) : 'NULL';
@@ -1259,7 +1250,7 @@ class GFAPI {
 
 		foreach ( $form['fields'] as $field ) {
 			/* @var GF_Field $field */
-			if ( in_array( $field->type, array( 'html', 'page', 'section' ) ) ) {
+			if ( $field->displayOnly ) {
 				continue;
 			}
 			self::queue_batch_field_operation( $form, $entry, $field );
@@ -1418,6 +1409,173 @@ class GFAPI {
 				$result = GFFormsModel::update_entry_field_value( $form, $entry, $field, $lead_detail_id, $input_id, $value, $item_index );
 			}
 		}
+
+		return $result;
+	}
+
+	// ENTRY NOTES ------------------------------------------------
+
+	/**
+	 * Get notes based on search criteria.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param array $search_criteria Array of search criteria.
+	 * @param array $sorting Sort key and direction.
+	 * @return array|bool
+	 */
+	public static function get_notes( $search_criteria = array(), $sorting = null ) {
+
+		if ( ! $sorting ) {
+			$sorting = array(
+				'key'        => 'id',
+				'direction'  => 'ASC',
+				'is_numeric' => true,
+			);
+		}
+
+		$notes = GFFormsModel::get_notes( $search_criteria, $sorting );
+
+		if ( empty( $notes ) ) {
+			return false;
+		}
+
+		return $notes;
+	}
+
+	/**
+	 * Get note by ID.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param int $note_id ID of the note to retrieve.
+	 * @return array|WP_Error
+	 */
+	public static function get_note( $note_id ) {
+		$note = GFFormsModel::get_notes( array( 'id' => $note_id ) );
+
+		if ( empty( $note ) ) {
+			return new WP_Error( 'note_not_found', __( 'Note not found', 'gravityforms' ) );
+		}
+
+		return $note[0];
+	}
+
+	/**
+	 * Create one note for an entry.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param int    $entry_id ID of the entry to add the note to.
+	 * @param int    $user_id ID of the user to associate with the note.
+	 * @param string $user_name Name of the user to associate with the note.
+	 * @param string $note Text of the note.
+	 * @param string $note_type Note type.
+	 * @param null   $sub_type Not sub-type.
+	 * @return array|int|void|WP_Error
+	 */
+	public static function add_note( $entry_id, $user_id, $user_name, $note, $note_type = 'user', $sub_type = null ) {
+		if ( gf_upgrade()->get_submissions_block() ) {
+			return new WP_Error( 'submissions_blocked', __( 'Submissions are currently blocked due to an upgrade in progress', 'gravityforms' ) );
+		}
+
+		if ( ! self::entry_exists( $entry_id ) ) {
+			return new WP_Error( 'invalid_entry', __( 'Invalid entry', 'gravityforms' ), $entry_id );
+		}
+
+		if ( empty( $note ) || ! is_string( $note ) ) {
+			return new WP_Error( 'invalid_note', __( 'Invalid or empty note', 'gravityforms' ), $entry_id );
+		}
+
+		$new_note = GFFormsModel::add_note( intval( $entry_id ), $user_id, $user_name, wp_kses_post( $note ), sanitize_text_field( $note_type ), sanitize_text_field( $sub_type ) );
+
+		return $new_note;
+	}
+
+	/**
+	 * Delete one note.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param int $note_id ID of the note to delete.
+	 * @return int|WP_Error ID of the deleted note.
+	 */
+	public static function delete_note( $note_id ) {
+		$result = GFFormsModel::delete_note( $note_id );
+
+		if ( ! $result ) {
+			return new WP_Error( 'invalid_note', __( 'Invalid note', 'gravityforms' ), $note_id );
+		}
+
+		return $result;
+	}
+
+	/**
+	 * Update a note.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param array $note {
+	 * 		Note data to update.
+	 *
+	 *		@type int    $entry_id     ID of the entry associated with the note.
+	 *		@type int    $user_id      ID of the user associated with the note.
+	 * 		@type string $user_name    Name of the user associated with the note.
+	 *		@type string $date_created Date and time the note was created, in SQL datetime format.
+	 *		@type string $value        The text of the note.
+	 *		@type string $note_type    The note type.
+	 *		@type string $sub_type     The note subtype.
+	 * }
+	 * @param int   $note_id ID of the note to update.
+	 * @return bool|WP_Error
+	 */
+	public static function update_note( $note, $note_id = '' ) {
+		if ( gf_upgrade()->get_submissions_block() ) {
+			return new WP_Error( 'submissions_blocked', __( 'Submissions are currently blocked due to an upgrade in progress', 'gravityforms' ) );
+		}
+
+		if ( ! is_array( $note ) || empty( $note ) ) {
+			return new WP_Error( 'invalid_note_format', __( 'Invalid note format', 'gravityforms' ) );
+		}
+
+		if ( empty( $note_id ) ) {
+			if ( rgar( $note, 'id' ) ) {
+				$note_id = absint( $note['id'] );
+			}
+		} else {
+			$note_id = absint( $note_id );
+		}
+
+		if ( empty( $note_id ) ) {
+			return new WP_Error( 'missing_note_id', __( 'Missing note id', 'gravityforms' ) );
+		}
+
+		// make sure the note exists.
+		$current_note = self::get_note( $note_id );
+		if ( ! $current_note || is_wp_error( $current_note ) ) {
+			return new WP_Error( 'note_not_found', __( 'Note not found', 'gravityforms' ) );
+		}
+
+		$note_properties = array(
+			'id',
+			'entry_id',
+			'user_id',
+			'user_name',
+			'date_created',
+			'value',
+			'note_type',
+			'sub_type',
+		);
+
+		$current_note_array = (array) $current_note;
+
+		foreach ( $note_properties as $property ) {
+			if ( ! isset( $note[ $property ] ) ) {
+				$note[ $property ] = $current_note_array[ $property ];
+			}
+		}
+
+		$result = GFFormsModel::update_note( $note['id'], $note['entry_id'], $note['user_id'], $note['user_name'], $note['date_created'], $note['value'], $note['note_type'], $note['sub_type'] );
 
 		return $result;
 	}
@@ -1606,17 +1764,22 @@ class GFAPI {
 	 * @access public
 	 * @global $wpdb
 	 *
-	 * @param mixed  $feed_ids   The ID of the Feed or an array of Feed IDs.
-	 * @param int    $form_id    The ID of the Form to which the Feeds belong.
-	 * @param string $addon_slug The slug of the add-on to which the Feeds belong.
-	 * @param bool   $is_active  If the feed is active.
+	 * @param mixed       $feed_ids   The ID of the Feed or an array of Feed IDs.
+	 * @param null|int    $form_id    The ID of the Form to which the Feeds belong.
+	 * @param null|string $addon_slug The slug of the add-on to which the Feeds belong.
+	 * @param bool        $is_active  If the feed is active.
 	 *
 	 * @return array|WP_Error Either an array of Feed objects or a WP_Error instance.
 	 */
 	public static function get_feeds( $feed_ids = null, $form_id = null, $addon_slug = null, $is_active = true ) {
 		global $wpdb;
 
-		$table       = $wpdb->prefix . 'gf_addon_feed';
+		$table = $wpdb->prefix . 'gf_addon_feed';
+
+		if ( ! GFCommon::table_exists( $table ) ) {
+			return self::get_missing_table_wp_error( $table );
+		}
+
 		$where_arr   = array();
 		$where_arr[] = $wpdb->prepare( 'is_active=%d', $is_active );
 		if ( false === empty( $form_id ) ) {
@@ -1630,7 +1793,7 @@ class GFAPI {
 				$feed_ids = array( $feed_ids );
 			}
 			$in_str_arr  = array_fill( 0, count( $feed_ids ), '%d' );
-			$in_str      = join( $in_str_arr, ',' );
+			$in_str      = join( ',', $in_str_arr );
 			$where_arr[] = $wpdb->prepare( "id IN ($in_str)", $feed_ids );
 		}
 
@@ -1670,6 +1833,10 @@ class GFAPI {
 		}
 
 		$table = $wpdb->prefix . 'gf_addon_feed';
+
+		if ( ! GFCommon::table_exists( $table ) ) {
+			return self::get_missing_table_wp_error( $table );
+		}
 
 		$sql = $wpdb->prepare( "DELETE FROM {$table} WHERE id=%d", $feed_id );
 
@@ -1744,7 +1911,12 @@ class GFAPI {
 			return new WP_Error( 'submissions_blocked', __( 'Submissions are currently blocked due to an upgrade in progress', 'gravityforms' ) );
 		}
 
-		$table          = $wpdb->prefix . 'gf_addon_feed';
+		$table = $wpdb->prefix . 'gf_addon_feed';
+
+		if ( ! GFCommon::table_exists( $table ) ) {
+			return self::get_missing_table_wp_error( $table );
+		}
+
 		$feed_meta_json = json_encode( $feed_meta );
 		$sql            = $wpdb->prepare( "INSERT INTO {$table} (form_id, meta, addon_slug) VALUES (%d, %s, %s)", $form_id, $feed_meta_json, $addon_slug );
 
@@ -1755,6 +1927,19 @@ class GFAPI {
 		}
 
 		return $wpdb->insert_id;
+	}
+
+	/**
+	 * Returns the missing_table WP_Error.
+	 *
+	 * @since 2.4.22
+	 *
+	 * @param string $table The name of the table which does not exist.
+	 *
+	 * @return WP_Error
+	 */
+	private static function get_missing_table_wp_error( $table ) {
+		return new WP_Error( 'missing_table', sprintf( __( 'The %s table does not exist.', 'gravityforms' ), $table ) );
 	}
 
 	// NOTIFICATIONS ----------------------------------------------
